@@ -80,6 +80,7 @@
     "iced-runtime": "1.0.3",
     "immutable": "3.8.1",
     "jsonfile": "3.0.0",
+    "klaw-sync": "2.1.0",
     "lodash": "4.17.4",
     "material-ui": "0.15.4",
     "menubar": "5.2.3",
@@ -87,6 +88,7 @@
     "moment": "2.18.1",
     "path-parse": "1.0.5",
     "prettier": "1.3.1",
+    "prop-types": "15.5.10",
     "purepack": "1.0.4",
     "qrcode-generator": "1.2.0",
     "re-reselect": "0.5.0",
@@ -133,9 +135,9 @@
     "css-loader": "0.28.4",
     "del": "2.2.2",
     "devtron": "1.4.0",
-    "electron": "1.4.12",
+    "electron": "1.6.10",
     "electron-download": "git://github.com/keybase/electron-download#5350b5a1bad8ba7bb8c7201abe59ae3a03c3d75d",
-    "electron-packager": "8.3.0",
+    "electron-packager": "8.7.0",
     "eslint": "3.19.0",
     "eslint-config-standard": "10.2.1",
     "eslint-config-standard-jsx": "4.0.1",
